moved edittasks to new package
