+++ src/main/java/cubicchunks/converter/lib/conf/command/commands/SetCommand.java
@@ -6,9 +6,9 @@ import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.builder.RequiredArgumentBuilder;
 import cubicchunks.converter.lib.conf.command.EditTaskContext;
 import cubicchunks.converter.lib.conf.command.arguments.BoundingBoxArgument;
-import cubicchunks.converter.lib.util.BlockEditTask;
+import cubicchunks.converter.lib.util.edittask.BlockEditTask;
 import cubicchunks.converter.lib.util.BoundingBox;
-import cubicchunks.converter.lib.util.EditTask;
+import cubicchunks.converter.lib.util.edittask.EditTask;
 
 public class SetCommand {
     public static void register(CommandDispatcher<EditTaskContext> dispatcher) {
