+++ src/main/java/cubicchunks/converter/lib/convert/io/DualSourceCubicChunkReader.java
@@ -28,12 +28,11 @@ import static cubicchunks.converter.lib.util.Utils.interruptibleConsumer;
 import com.carrotsearch.hppc.IntArrayList;
 import com.carrotsearch.hppc.cursors.IntCursor;
 import cubicchunks.converter.lib.Dimension;
-import cubicchunks.converter.lib.Dimensions;
 import cubicchunks.converter.lib.conf.ConverterConfig;
-import cubicchunks.converter.lib.convert.data.CubicChunksColumnData;
 import cubicchunks.converter.lib.convert.data.DualSourceCubicChunksColumnData;
 import cubicchunks.converter.lib.convert.impl.DualSourceSaveCubeColumns;
 import cubicchunks.converter.lib.util.*;
+import cubicchunks.converter.lib.util.edittask.EditTask;
 import cubicchunks.regionlib.api.region.IRegionProvider;
 import cubicchunks.regionlib.api.region.key.RegionKey;
 import cubicchunks.regionlib.impl.EntryLocation2D;
